requirements.txt
bcrypt==4.0.1
Pillow==9.3.0
tk==0.1.0

